Change default program stack from max 16M to min 16M
I think the original setting was a mistake. I think a 16M program stack
for Lace tasks is reasonable. The default in Linux is 8M. Depending on
the program (or the input for the program), a larger program stack is
needed, since recursive programs may need a large program stack if there
are many recursive calls...