Improve benchmark script
The new script uses bootstrapping to compute the standard error of the median.
The script runs each benchmark at least 10 times, and then until the standard error
of the median is less than 0.1%. The reported median is the mean of the bootstrap median.

The logic behind this is that benchmarks may be influenced by rare spiky noise.
By using the median rather than the mean, we reduce the influence of this spiky noise.
In order to know when to stop benchmarking, we need a stopping condition based on the
median, and using bootstrapping is one method to obtain a standard error of the median.

We use standard parameters for the bootstrapping, i.e., B=1000 samples with replacement,
and we sample as many values as in the original data.